Move argument parsing out from main.py
Moved argument parsing out from main.py in order to prevent circular
import issues when using it outside of main.py.

Ticket: None
Changelog: None
